Move pattern view components from design/config/view to design/form/components, to be closer to where they are consumed.
