+++ packages/design/src/Form/components/index.tsx
@@ -1,3 +1,5 @@
+import { type ComponentForPattern } from '..';
+
 import Address from './Address';
 import Fieldset from './Fieldset';
 import FormSummary from './FormSummary';
@@ -5,7 +7,6 @@ import Paragraph from './Paragraph';
 import Sequence from './Sequence';
 import SubmissionConfirmation from './SubmissionConfirmation';
 import TextInput from './TextInput';
-import { type ComponentForPattern } from '../../Form';
 
 export const defaultPatternComponents: ComponentForPattern = {
   address: Address,
