+++ packages/design/src/index.ts
@@ -2,7 +2,7 @@ import '@uswds/uswds';
 
 export { default as AvailableFormList } from './AvailableFormList';
 export { default as Form } from './Form';
-export { default as FormRouter } from './FormRouter';
+export { defaultPatternComponents } from './Form/components';
 export { default as FormManager } from './FormManager';
 export { defaultPatternEditComponents } from './FormManager/FormEdit/components';
-export * from './config';
+export { default as FormRouter } from './FormRouter';
